changed volume to float

+++ app/src/main/java/controller/Controller.java
@@ -27,7 +27,7 @@ public class Controller {
     audioPlayer.playAudio();
   }
 
-  public void setMasterVolume(int volume){
+  public void setMasterVolume(float volume){
     audioPlayer.setVolume(volume);
   }
